extract silverbull/vet parsing to RankModel
