+++ common/src/main/java/com/wynntils/commands/WynntilsCommand.java
@@ -384,7 +384,8 @@ public class WynntilsCommand extends Command {
     }
 
     private int rescan(CommandContext<CommandSourceStack> context) {
-        Models.Character.scanCharacterInfo(true);
+        Models.Character.scanCharacterInfo();
+        Models.Rank.scanRankInfo(true);
         return 1;
     }
