+++ common/src/main/java/com/wynntils/features/inventory/HightlightDuplicateCosmeticsFeature.java
@@ -116,7 +116,7 @@ public class HightlightDuplicateCosmeticsFeature extends Feature {
         if (scrapMenu == null) return;
 
         // Silverbull subscribers have the first 2 slots filled automatically
-        List<Integer> selectedSlots = Models.Character.isSilverbullSubscriber()
+        List<Integer> selectedSlots = Models.Rank.isSilverbullSubscriber()
                 ? SELECTED_COSMETIC_SLOTS.subList(2, 5)
                 : SELECTED_COSMETIC_SLOTS;
