+++ common/src/main/java/com/wynntils/models/character/CharacterModel.java
@@ -50,25 +50,10 @@ public final class CharacterModel extends Model {
     private static final Pattern CHARACTER_ID_PATTERN = Pattern.compile("^[a-z0-9]{8}$");
     private static final Pattern INFO_MENU_CLASS_PATTERN = Pattern.compile("§7Class: §f(.+)");
     private static final Pattern INFO_MENU_LEVEL_PATTERN = Pattern.compile("§7Combat Lv: §f(\\d+)");
-    // Test in CharacterModel_SILVERBULL_PATTERN
-    private static final Pattern SILVERBULL_PATTERN = Pattern.compile("§7Subscription: §[ac][✖✔] ((?:Ina|A)ctive)");
-    // Test in CharacterModel_SILVERBULL_DURATION_PATTERN
-    private static final Pattern SILVERBULL_DURATION_PATTERN = Pattern.compile(
-            "§7Expiration: §f(?:(?<weeks>\\d+) weeks?)? ?(?:(?<days>\\d+) days?)? ?(?:(?<hours>\\d+) hours?)? ?(?:(?<minutes>\\d+) minutes?)? ?(?:(?<seconds>\\d+) seconds?)?");
-    // Test in CharacterModel_VETERAN_PATTERN
-    private static final Pattern VETERAN_PATTERN = Pattern.compile("§7Rank: §[6dba]Vet");
-    private static final Pattern SILVERBULL_JOIN_PATTERN =
-            Pattern.compile("§3Welcome to the §b✮ Silverbull Trading Company§3!");
-    private static final Pattern SILVERBULL_UPDATE_PATTERN = Pattern.compile("§7Your subscription has been extended.");
 
-    private static final int RANK_SUBSCRIPTION_INFO_SLOT = 0;
     public static final int CHARACTER_INFO_SLOT = 7;
     private static final int PROFESSION_INFO_SLOT = 17;
-    private static final int COSMETICS_SLOT = 25;
-    private static final int COSMETICS_BACK_SLOT = 9;
     public static final int GUILD_MENU_SLOT = 26;
-
-    public static final Component SILVERBULL_STAR = Component.literal(" ✮").withStyle(ChatFormatting.AQUA);
 
     private static final DeathScreenBar deathScreenBar = new DeathScreenBar();
 
@@ -83,15 +68,6 @@ public final class CharacterModel extends Model {
     private boolean reskinned;
     private int level;
 
-    @Persisted
-    private final Storage<Boolean> isVeteran = new Storage<>(false);
-
-    @Persisted
-    private final Storage<Long> silverbullExpiresAt = new Storage<>(0L);
-
-    @Persisted
-    private final Storage<ConfirmedBoolean> silverbullSubscriber = new Storage<>(ConfirmedBoolean.UNCONFIRMED);
-
     // A hopefully unique string for each character ("class"). This is part of the
     // full character uuid, as presented by Wynncraft in the tooltip.
     private String id = "-";
@@ -100,10 +76,6 @@ public final class CharacterModel extends Model {
         super(List.of());
 
         Handlers.BossBar.registerBar(deathScreenBar);
-    }
-
-    public boolean isSilverbullSubscriber() {
-        return silverbullSubscriber.get() == ConfirmedBoolean.TRUE;
     }
 
     public ClassType getClassType() {
@@ -136,10 +108,6 @@ public final class CharacterModel extends Model {
         return id;
     }
 
-    public boolean isVeteran() {
-        return isVeteran.get();
-    }
-
     // FIXME: Remove if this is not needed, or fix it for 2.1
     public boolean isHuntedMode() {
         return false;
@@ -168,25 +136,11 @@ public final class CharacterModel extends Model {
             updateCharacterId();
 
             // We need to scan character info and profession info as well.
-            scanCharacterInfo(e.isFirstJoinWorld());
+            scanCharacterInfo();
         }
     }
 
-    @SubscribeEvent
-    public void onChatReceived(ChatMessageReceivedEvent e) {
-        StyledText message = e.getOriginalStyledText();
 
-        StyledText trimmedMessage = message.trim();
-        if (trimmedMessage.matches(SILVERBULL_JOIN_PATTERN)) {
-            silverbullSubscriber.store(ConfirmedBoolean.TRUE);
-            return;
-        }
-
-        if (trimmedMessage.matches(SILVERBULL_UPDATE_PATTERN)) {
-            silverbullSubscriber.store(ConfirmedBoolean.TRUE);
-            return;
-        }
-    }
 
     @SubscribeEvent
     public void onBossBarAdd(BossBarAddedEvent event) {
@@ -218,7 +172,7 @@ public final class CharacterModel extends Model {
         }
     }
 
-    public void scanCharacterInfo(boolean forceParseEverything) {
+    public void scanCharacterInfo() {
         WynntilsMod.info("Scheduling character info query");
         QueryBuilder queryBuilder = ScriptedContainerQuery.builder("Character Info Query");
         queryBuilder.onError(msg -> WynntilsMod.warn("Error querying Character Info: " + msg));
@@ -227,22 +181,6 @@ public final class CharacterModel extends Model {
         queryBuilder.then(QueryStep.useItemInHotbar(InventoryUtils.COMPASS_SLOT_NUM)
                 .expectContainerTitle(ContainerModel.CHARACTER_INFO_NAME)
                 .processIncomingContainer(this::parseCharacterContainer));
-
-        if (forceParseEverything
-                || silverbullSubscriber.get() == ConfirmedBoolean.UNCONFIRMED
-                || (silverbullSubscriber.get() != ConfirmedBoolean.FALSE
-                        && System.currentTimeMillis() > silverbullExpiresAt.get())) {
-            // Open Cosmetics Menu
-            queryBuilder
-                    .then(QueryStep.clickOnSlot(COSMETICS_SLOT)
-                            .expectContainerTitle(ContainerModel.COSMETICS_MENU_NAME)
-                            .processIncomingContainer(this::parseCratesBombsCosmeticsContainer))
-                    .then(QueryStep.clickOnSlot(COSMETICS_BACK_SLOT)
-                            .expectContainerTitle(ContainerModel.CHARACTER_INFO_NAME));
-        } else {
-            WynntilsMod.info("Skipping silverbull subscription query ("
-                    + (silverbullExpiresAt.get() - System.currentTimeMillis()) + " ms left)");
-        }
 
         // Scan guild container, if the player is in a guild
         Models.Guild.addGuildContainerQuerySteps(queryBuilder);
@@ -264,48 +202,6 @@ public final class CharacterModel extends Model {
         WynntilsMod.info("Deducing character " + getCharacterString());
     }
 
-    private void parseCratesBombsCosmeticsContainer(ContainerContent container) {
-        ItemStack rankSubscriptionItem = container.items().get(RANK_SUBSCRIPTION_INFO_SLOT);
-
-        Matcher veteran = LoreUtils.matchLoreLine(rankSubscriptionItem, 0, VETERAN_PATTERN);
-
-        isVeteran.store(veteran.matches());
-
-        Matcher status = LoreUtils.matchLoreLine(rankSubscriptionItem, 0, SILVERBULL_PATTERN);
-        if (!status.matches()) {
-            WynntilsMod.warn("Could not parse Silverbull subscription status from item: "
-                    + LoreUtils.getLore(rankSubscriptionItem));
-            silverbullSubscriber.store(ConfirmedBoolean.FALSE);
-            return;
-        }
-
-        silverbullSubscriber.store(status.group(1).equals("Active") ? ConfirmedBoolean.TRUE : ConfirmedBoolean.FALSE);
-        if (silverbullSubscriber.get() != ConfirmedBoolean.TRUE) return;
-
-        Matcher expiry = LoreUtils.matchLoreLine(rankSubscriptionItem, 1, SILVERBULL_DURATION_PATTERN);
-        if (!expiry.matches()) {
-            WynntilsMod.warn("Could not parse Silverbull subscription expiry from item: "
-                    + LoreUtils.getLore(rankSubscriptionItem));
-            silverbullExpiresAt.store(0L);
-            return;
-        }
-        int weeks = expiry.group("weeks") == null ? 0 : Integer.parseInt(expiry.group("weeks"));
-        int days = expiry.group("days") == null ? 0 : Integer.parseInt(expiry.group("days"));
-        int hours = expiry.group("hours") == null ? 0 : Integer.parseInt(expiry.group("hours"));
-        int minutes = expiry.group("minutes") == null ? 0 : Integer.parseInt(expiry.group("minutes"));
-        int seconds = expiry.group("seconds") == null ? 0 : Integer.parseInt(expiry.group("seconds"));
-
-        long expiryTime = System.currentTimeMillis()
-                + TimeUnit.DAYS.toMillis(weeks * 7L + days)
-                + TimeUnit.HOURS.toMillis(hours)
-                + TimeUnit.MINUTES.toMillis(minutes)
-                + TimeUnit.SECONDS.toMillis(seconds);
-        silverbullExpiresAt.store(expiryTime);
-
-        WynntilsMod.info(
-                "Parsed Silverbull subscription status: " + silverbullSubscriber.get() + ", expires at: " + expiryTime);
-    }
-
     private void updateCharacterId() {
         ItemStack compassItem = McUtils.inventory().items.get(CHARACTER_INFO_SLOT);
         List<StyledText> compassLore = LoreUtils.getLore(compassItem);
@@ -325,8 +221,7 @@ public final class CharacterModel extends Model {
                 + classType + ", reskinned="
                 + reskinned + ", level="
                 + level + ", id="
-                + id + ", silverbullSubscriber="
-                + silverbullSubscriber.get() + '}';
+                + id + '}';
     }
 
     private void parseCharacterFromCharacterMenu(ItemStack characterInfoItem) {
