+++ fabric/src/test/java/TestRegex.java
@@ -12,6 +12,7 @@ import com.wynntils.handlers.chat.type.RecipientType;
 import com.wynntils.models.abilities.ShamanTotemModel;
 import com.wynntils.models.abilities.bossbars.OphanimBar;
 import com.wynntils.models.character.CharacterModel;
+import com.wynntils.models.character.RankModel;
 import com.wynntils.models.combat.CombatModel;
 import com.wynntils.models.combat.label.DamageLabelParser;
 import com.wynntils.models.combat.label.KillLabelParser;
@@ -157,35 +158,6 @@ public class TestRegex {
         p.shouldMatch("§6 - §a✔ §f24§7²");
         p.shouldMatch("§6 - §a✔ §f648§7²");
         p.shouldMatch("§6 - §c✖ §f24§7²");
-    }
-
-    @Test
-    public void CharacterModel_SILVERBULL_PATTERN() {
-        PatternTester p = new PatternTester(CharacterModel.class, "SILVERBULL_PATTERN");
-        p.shouldMatch("§7Subscription: §c✖ Inactive");
-        p.shouldMatch("§7Subscription: §a✔ Active");
-    }
-
-    @Test
-    public void CharacterModel_SILVERBULL_DURATION_PATTERN() {
-        PatternTester p = new PatternTester(CharacterModel.class, "SILVERBULL_DURATION_PATTERN");
-        p.shouldMatch("§7Expiration: §f1 week 5 days");
-        p.shouldMatch("§7Expiration: §f5 days");
-        p.shouldMatch("§7Expiration: §f1 week");
-        p.shouldMatch("§7Expiration: §f2 days 12 hours");
-    }
-
-    @Test
-    public void CharacterModel_VETERAN_PATTERN() {
-        PatternTester p = new PatternTester(CharacterModel.class, "VETERAN_PATTERN");
-        // Champion
-        p.shouldMatch("§7Rank: §6Vet");
-        // Hero
-        p.shouldMatch("§7Rank: §dVet");
-        // VIP+
-        p.shouldMatch("§7Rank: §bVet");
-        // VIP
-        p.shouldMatch("§7Rank: §aVet");
     }
 
     @Test
@@ -910,6 +882,35 @@ public class TestRegex {
         p.shouldNotMatch("§f⬡ §7Mobs Killed: §f");
         p.shouldMatch("§f⬡ §7Wars Won: §f164");
         p.shouldMatch("§f⬡ §7Raids Won: §f0");
+    }
+
+    @Test
+    public void RankModel_SILVERBULL_PATTERN() {
+        PatternTester p = new PatternTester(RankModel.class, "SILVERBULL_PATTERN");
+        p.shouldMatch("§7Subscription: §c✖ Inactive");
+        p.shouldMatch("§7Subscription: §a✔ Active");
+    }
+
+    @Test
+    public void RankModel_SILVERBULL_DURATION_PATTERN() {
+        PatternTester p = new PatternTester(RankModel.class, "SILVERBULL_DURATION_PATTERN");
+        p.shouldMatch("§7Expiration: §f1 week 5 days");
+        p.shouldMatch("§7Expiration: §f5 days");
+        p.shouldMatch("§7Expiration: §f1 week");
+        p.shouldMatch("§7Expiration: §f2 days 12 hours");
+    }
+
+    @Test
+    public void RankModel_VETERAN_PATTERN() {
+        PatternTester p = new PatternTester(RankModel.class, "VETERAN_PATTERN");
+        // Champion
+        p.shouldMatch("§7Rank: §6Vet");
+        // Hero
+        p.shouldMatch("§7Rank: §dVet");
+        // VIP+
+        p.shouldMatch("§7Rank: §bVet");
+        // VIP
+        p.shouldMatch("§7Rank: §aVet");
     }
 
     @Test
